Refactor output parsing and format instruction logic
Extracting the logic into separate OutputParser classes, adding List and Set parsers, and creating a dedicated resolver to handle parser resolution.
